requirements.txt
Django>=3.2,<4.0
django-allauth>=0.42.0 
django-crispy-forms>=1.9.1 
django-debug-toolbar>=2.2 
django-environ>=0.4.5 
django-model-utils>=4.0.0 
django-redis>=4.12.1 
djangorestframework>=3.11.0 
python-decouple>=3.3 
python-slugify>=4.0.1 
requests>=2.24.0 
crispy-bootstrap4>=0.1
whitenoise>=6.0.0
psycopg2-binary>=2.9.0
dj-database-url>=1.0.0
django-storages>=1.13.2
boto3>=1.26.0